hybrid-retriever-modular-mcp/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
qdrant-client>=1.7.0
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
chardet>=5.2.0
# Korean morpheme tokenization for FTS5 keyword search.
kiwipiepy>=0.17.0
# Embedded property-graph DB (Cypher) used by the graph_query tool.
kuzu>=0.4.0
# Used directly by retriever/embedding_client.py for InsecureRequestWarning.
# Normally pulled transitively by `requests`; listed for explicitness.
urllib3>=1.26.0

# Required on Windows by portalocker (used by qdrant-client)
pywin32>=306

# Modular RAG: Haystack 2.x pipelines and Hypster configuration space.
# These power the retriever/ package; legacy scripts/ modules now delegate
# to them while preserving their public surface.
haystack-ai>=2.10.0
hypster>=0.3.0
